Adds -L flag to silence sysLog messages.
A quick note on intent here: Critical alerts often come from systems which are neglected, or are otherwise not well maintained- and far too often, tracking down "which host is the one which sent that" becomes a real problem.

With that, my preference as the author, for utilities like this, is to lazy-dump messages to syslog as a default.

I've gone back/forth on breaking the "rule of no surprise" here, in favor of hopefully saving somebody some time- in the middle of some hard night, filled with breaking systems.

Signed-off-by: Isaac (.ike) Levy <isaac.levy@enigma.io>